requirements.txt
# --- Core AI & Agent Framework ---
# google-adk==1.8.0 # Assuming this might be internal or placeholder, keeping for now or replace with specific libs if needed.
google-cloud-aiplatform==1.38.1 # Updated to ensure vertexai is included
google-genai>=0.3.0

# --- Web Framework & API ---
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
requests>=2.31.0
aiohttp>=3.9.3

# --- Data & Storage ---
google-cloud-storage>=2.14.0
google-cloud-firestore>=2.15.0
google-cloud-discoveryengine>=0.11.0
pydantic>=2.6.0
pydantic-settings>=2.2.0

# --- Utilities & Processing ---
python-dotenv>=1.0.1
msal>=1.26.0
beautifulsoup4>=4.12.0
markdownify>=0.11.6
markdown
bleach
jsonschema>=4.19.0
PyYAML>=6.0.1
